+++ .travis.yml
@@ -15,9 +15,9 @@ script:
   - xctool -project Swift/Adhan.xcodeproj -scheme AdhanTests -sdk iphonesimulator build test
   - JavaScript/Tests/runtests
   - node-qunit-phantomjs JavaScript/Tests/tests.html
-  - cd $TRAVIS_BUILD_DIR/Python && tox --hashseed 0
+  - cd $TRAVIS_BUILD_DIR/Python && make test
 after_success:
-  - cd $TRAVIS_BUILD_DIR/Python && tox -e coverage-report
+  - cd $TRAVIS_BUILD_DIR/Python && make coverage-report
   - codecov
 notifications:
   email: false
