+++ Python/MANIFEST.in
@@ -1,3 +1,3 @@
-include *.rst *.txt LICENSE tox.ini docs/Makefile .coveragerc
+include *.rst *.txt LICENSE tox.ini requirements.txt Makefile docs/Makefile .coveragerc
 recursive-include tests *.py
 recursive-include tests *.json
