+++ Python/Makefile
@@ -1,0 +1,8 @@
+init:
+	pip install -r requirements.txt
+
+test:
+	tox --hashseed 0
+
+coverage-report:
+	tox -e coverage-report
