+++ Python/tox.ini
@@ -16,7 +16,7 @@ commands = coverage run --parallel -m pytest {posargs}
 [testenv:flake8]
 basepython = python3.5
 deps = flake8
-commands = flake8 src tests setup.py
+commands = flake8 pyadhan tests setup.py
 
 
 [testenv:docs]
